requirements.txt
ipywidgets==8.1.5
jupyter-book==1.0.3
jupyter==1.1.1
jupyterlab==4.2.5
numpy==2.2.2
lazypredict-nightly==0.3.2
lightgbm==4.5.0
pandas==2.2.3
pip==24.3.1
plotly==5.24.1
prettyprinter==0.18.0
pytest==8.3.4
requests==2.32.3
scikit-learn==1.6.1
scikit-optimize==0.10.2
seaborn==0.13.2
shapely==2.0.6
statsmodels==0.14.4
streamlit==1.41.1
xgboost==2.1.3
